Develop

+++ src/modules/podcast/services/RemovePodcast.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import PodcastRepository from '../infra/prisma/repositories/PodcastRepository';
 
 @Injectable()
@@ -9,6 +9,9 @@ export default class RemovePodcastService {
   ) {}
 
   public async execute(podcastId: string) {
+    const podcastExist = await this.podcastRepository.showPodcast(podcastId);
+    if (!podcastExist) throw new NotFoundException('Podcast not found');
+
     const podcast = await this.podcastRepository.deletePodcast(podcastId);
     return podcast;
   }
